# RequestManagementDB/RequestManagementDB-api/src/main/java/RequestManagementDB/model/EmrgcyRequestModel.java
/**
 * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

package RequestManagementDB.model;

import com.liferay.portal.kernel.bean.AutoEscape;
import com.liferay.portal.kernel.model.BaseModel;
import com.liferay.portal.kernel.model.ShardedModel;
import com.liferay.portal.kernel.model.StagedAuditedModel;

import java.util.Date;

import org.osgi.annotation.versioning.ProviderType;

/**
 * The base model interface for the EmrgcyRequest service. Represents a row in the &quot;Fire_EmrgcyRequest&quot; database table, with each column mapped to a property of this class.
 *
 * <p>
 * This interface and its corresponding implementation <code>RequestManagementDB.model.impl.EmrgcyRequestModelImpl</code> exist only as a container for the default property accessors generated by ServiceBuilder. Helper methods and all application logic should be put in <code>RequestManagementDB.model.impl.EmrgcyRequestImpl</code>.
 * </p>
 *
 * @author Brian Wing Shun Chan
 * @see EmrgcyRequest
 * @generated
 */
@ProviderType
public interface EmrgcyRequestModel
	extends BaseModel<EmrgcyRequest>, ShardedModel, StagedAuditedModel {

	/*
	 * NOTE FOR DEVELOPERS:
	 *
	 * Never modify or reference this interface directly. All methods that expect a emrgcy request model instance should use the {@link EmrgcyRequest} interface instead.
	 */

	/**
	 * Returns the primary key of this emrgcy request.
	 *
	 * @return the primary key of this emrgcy request
	 */
	public long getPrimaryKey();

	/**
	 * Sets the primary key of this emrgcy request.
	 *
	 * @param primaryKey the primary key of this emrgcy request
	 */
	public void setPrimaryKey(long primaryKey);

	/**
	 * Returns the uuid of this emrgcy request.
	 *
	 * @return the uuid of this emrgcy request
	 */
	@AutoEscape
	@Override
	public String getUuid();

	/**
	 * Sets the uuid of this emrgcy request.
	 *
	 * @param uuid the uuid of this emrgcy request
	 */
	@Override
	public void setUuid(String uuid);

	/**
	 * Returns the request ID of this emrgcy request.
	 *
	 * @return the request ID of this emrgcy request
	 */
	public long getRequestId();

	/**
	 * Sets the request ID of this emrgcy request.
	 *
	 * @param requestId the request ID of this emrgcy request
	 */
	public void setRequestId(long requestId);

	/**
	 * Returns the company ID of this emrgcy request.
	 *
	 * @return the company ID of this emrgcy request
	 */
	@Override
	public long getCompanyId();

	/**
	 * Sets the company ID of this emrgcy request.
	 *
	 * @param companyId the company ID of this emrgcy request
	 */
	@Override
	public void setCompanyId(long companyId);

	/**
	 * Returns the user ID of this emrgcy request.
	 *
	 * @return the user ID of this emrgcy request
	 */
	@Override
	public long getUserId();

	/**
	 * Sets the user ID of this emrgcy request.
	 *
	 * @param userId the user ID of this emrgcy request
	 */
	@Override
	public void setUserId(long userId);

	/**
	 * Returns the user uuid of this emrgcy request.
	 *
	 * @return the user uuid of this emrgcy request
	 */
	@Override
	public String getUserUuid();

	/**
	 * Sets the user uuid of this emrgcy request.
	 *
	 * @param userUuid the user uuid of this emrgcy request
	 */
	@Override
	public void setUserUuid(String userUuid);

	/**
	 * Returns the user name of this emrgcy request.
	 *
	 * @return the user name of this emrgcy request
	 */
	@AutoEscape
	@Override
	public String getUserName();

	/**
	 * Sets the user name of this emrgcy request.
	 *
	 * @param userName the user name of this emrgcy request
	 */
	@Override
	public void setUserName(String userName);

	/**
	 * Returns the create date of this emrgcy request.
	 *
	 * @return the create date of this emrgcy request
	 */
	@Override
	public Date getCreateDate();

	/**
	 * Sets the create date of this emrgcy request.
	 *
	 * @param createDate the create date of this emrgcy request
	 */
	@Override
	public void setCreateDate(Date createDate);

	/**
	 * Returns the modified date of this emrgcy request.
	 *
	 * @return the modified date of this emrgcy request
	 */
	@Override
	public Date getModifiedDate();

	/**
	 * Sets the modified date of this emrgcy request.
	 *
	 * @param modifiedDate the modified date of this emrgcy request
	 */
	@Override
	public void setModifiedDate(Date modifiedDate);

	/**
	 * Returns the title of this emrgcy request.
	 *
	 * @return the title of this emrgcy request
	 */
	@AutoEscape
	public String getTitle();

	/**
	 * Sets the title of this emrgcy request.
	 *
	 * @param title the title of this emrgcy request
	 */
	public void setTitle(String title);

	/**
	 * Returns the description of this emrgcy request.
	 *
	 * @return the description of this emrgcy request
	 */
	@AutoEscape
	public String getDescription();

	/**
	 * Sets the description of this emrgcy request.
	 *
	 * @param description the description of this emrgcy request
	 */
	public void setDescription(String description);

	/**
	 * Returns the location of this emrgcy request.
	 *
	 * @return the location of this emrgcy request
	 */
	@AutoEscape
	public String getLocation();

	/**
	 * Sets the location of this emrgcy request.
	 *
	 * @param location the location of this emrgcy request
	 */
	public void setLocation(String location);

	/**
	 * Returns the status of this emrgcy request.
	 *
	 * @return the status of this emrgcy request
	 */
	@AutoEscape
	public String getStatus();

	/**
	 * Sets the status of this emrgcy request.
	 *
	 * @param status the status of this emrgcy request
	 */
	public void setStatus(String status);

	/**
	 * Returns the assigned team of this emrgcy request.
	 *
	 * @return the assigned team of this emrgcy request
	 */
	@AutoEscape
	public String getAssignedTeam();

	/**
	 * Sets the assigned team of this emrgcy request.
	 *
	 * @param assignedTeam the assigned team of this emrgcy request
	 */
	public void setAssignedTeam(String assignedTeam);

	/**
	 * Returns the requester user ID of this emrgcy request.
	 *
	 * @return the requester user ID of this emrgcy request
	 */
	public long getRequesterUserId();

	/**
	 * Sets the requester user ID of this emrgcy request.
	 *
	 * @param requesterUserId the requester user ID of this emrgcy request
	 */
	public void setRequesterUserId(long requesterUserId);

	/**
	 * Returns the requester user uuid of this emrgcy request.
	 *
	 * @return the requester user uuid of this emrgcy request
	 */
	public String getRequesterUserUuid();

	/**
	 * Sets the requester user uuid of this emrgcy request.
	 *
	 * @param requesterUserUuid the requester user uuid of this emrgcy request
	 */
	public void setRequesterUserUuid(String requesterUserUuid);

	@Override
	public EmrgcyRequest cloneWithOriginalValues();

	public default String toXmlString() {
		return null;
	}

}